\name{n.obs.mot}
\alias{n.obs.mot}
\title{Number of observed motifs}
\usage{
n.obs.mot(motif, seq, overlap = TRUE)
}
\arguments{
\item{motif}{A motif in the form of a text string. Can be a regular expression.}

\item{seq}{a 'seq' object or a string or a list of 'seq' objects or strings}

\item{overlap}{Logical, should overlapping motifs be allowed.}
}
\description{
Function returning (as numeric) the number of observed motifs in a single sequence or a list of sequences.
}
\seealso{
\code{\link{n.obs.mot}}
}
